Add syscall tracing capabilities
This script allows tracing of system calls, providing insights into
process behavior and system interactions. It leverages kprobes to
monitor syscall entry and exit points.

This script can be used as a preliminary analysis of the potential root
cause of the syscall performance latency. Then, we can use the function
tracer, such as trace-cmd, to precisely trace the syscall function
graph.

The syscall map written in the script is generated by the following
command:

awk '/__NR_/ { gsub("__NR_", "", $2); print "@syscall_map[" $3 "] = \"" $2 "\";" }' \
/usr/include/x86_64-linux-gnu/asm/unistd_64.h > syscall_map.bt

This command searches for lines containing __NR_ in the unistd_64.h
header file, which defines syscall numbers.

It removes the __NR_ prefix from the syscall name and formats each line
as a bpftrace map assignment.

Signed-off-by: Gavin Guo <gavinguo@igalia.com>